Use package extension to eliminate CUDA from core dependencies
