+++ .github/workflows/ci.yml
@@ -86,6 +86,8 @@ jobs:
             Pkg.develop(path="src/ReinforcementLearningBase")
             Pkg.develop(path="src/ReinforcementLearningCore")
             Pkg.develop(path="src/ReinforcementLearningEnvironments")
+            Pkg.test("ReinforcementLearningCore", coverage=true)
+            Pkg.add("CUDA") # Retest with CUDA extension loaded
             Pkg.test("ReinforcementLearningCore", coverage=true)'
       - uses: julia-actions/julia-processcoverage@v1
       - uses: codecov/codecov-action@v3
